+++ annotation/lib/dio.dart
@@ -30,3 +30,9 @@ class DioResponseType {
   final ResponseType responseType;
   const DioResponseType(this.responseType);
 }
+
+class HttpResponse<T> {
+  final T data;
+  final Response response;
+  HttpResponse(this.data, this.response);
+}